% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variousOtherHelperFunctions.R
\name{addStationCameraID}
\alias{addStationCameraID}
\title{Add station camera ID}
\usage{
addStationCameraID(
  intable,
  dirs_short,
  stationCol,
  cameraCol,
  cameraID,
  hasStationFolders,
  i,
  IDfrom,
  stationIDposition,
  cameraIDposition = NULL
)
}
\arguments{
\item{stationCol}{character. Name of the camera trap station column. Assumes
"Station" if undefined.}

\item{cameraID}{This argument is deprecated. Use \code{cameraIDfrom}.}

\item{IDfrom}{This argument is deprecated. Use \code{speciesIDfrom}.}

\item{stationIDposition}{integer. The numeric position within the
filepath/directory structure of the 'Station' directories (e.g.
"C:/Documents/CamProject/Region/Site/Transect/Station/Camera/Species/...",
the 'Station' directory position = 6).Only need to use this argument when
information from one or more directories (i.e. folders) before the 'Station'
level directory are to be used to populate columns within the record table
(e.g. "SurveyPeriod/Region/Site/Transect/Station/..."). If this argument is
missing, the function assumes the 'Station' directory is next in the
filepath as specified in \code{inDir}.}

\item{cameraIDposition}{integer. The numeric position of the 'Camera'
directories within the filepath/directory structure (e.g.
"C:/Documents/CamProject/Region/Site/Transect/Station/Camera/Species,...",
the camera ID directory position = 7). Only need to use this argument when
\code{cameraIDfrom} = "directory" and there are one or more directories
(i.e. folders) after the 'Species' level directory (e.g.
"SurveyPeriod/Region/Site/Transect/Station/Species/Counts/...").}
}
\description{
Add station camera ID
}
