% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordTableIndividual.R
\name{recordTableIndividualAssist}
\alias{recordTableIndividualAssist}
\title{Generate a single-species record table with individual identification from
camera trap images}
\usage{
recordTableIndividualAssist(
  inDir,
  cameraID,
  hasStationFolders,
  IDfrom,
  StationIDfrom = "directory",
  speciesIDfrom,
  cameraIDfrom,
  camerasIndependent,
  individualIDfrom,
  exclude,
  minDeltaTime = 0,
  deltaTimeComparedTo,
  timeZone,
  stationCol,
  writecsv = FALSE,
  outDir,
  metadataHierarchyDelimitor = "|",
  metadataSpeciesTag,
  metadataIDTag,
  additionalMetadataTags,
  removeDuplicateRecords = TRUE,
  stationIDposition = NULL,
  speciesPosition = NULL,
  cameraIDposition = NULL,
  directoryInfoPositions,
  directoryInfoNames
)
}
\arguments{
\item{inDir}{character. Directory filepath containing 'Station' directories.
It must either contain images in species subdirectories (e.g.
"inDir/StationA/SpeciesA") or images with species metadata tags instead of
'Species' directories (e.g. "inDir/StationA"). Only include path up to
root/parent folder in which the 'Station'/'Site' subfolders are located
(e.g. "C:/Documents/CamProject" ...). 'Station' folders would be next in the
file path but are not included when defining \code{inDir}.}

\item{cameraID}{This argument is deprecated. Use \code{cameraIDfrom}.}

\item{hasStationFolders}{Logical. Whether station column is taken from directory}

\item{IDfrom}{This argument is deprecated. Use \code{speciesIDfrom}.}

\item{StationIDfrom}{character. Read station ID from image "filename" or
station "directory" names. "filename" requires images renamed with
\code{\link{imageRename}}.}

\item{speciesIDfrom}{character. Read species ID from image "metadata" or from
species "directory" names.}

\item{cameraIDfrom}{character. Read camera ID from image "filename" or camera
"directory" names. "filename" requires images renamed with
\code{\link{imageRename}}. "directory" requires a camera subdirectory within
'Station' directories (e.g. "Station/Camera/...").}

\item{camerasIndependent}{logical. If \code{TRUE}, species records are
considered independent between cameras at same station.}

\item{individualIDfrom}{character. Read individual ID from image metadata
("metadata") of from directory names ("directory")}

\item{exclude}{character. Vector of species names to be excluded from the
record table (e.g. c("Dog", "Cat", ...)).}

\item{minDeltaTime}{integer. The minimum specified time difference (in
minutes) between records of the same species, at the same station, that are
to be considered independent.}

\item{deltaTimeComparedTo}{character. For two records to be considered
independent, the second record must be at least \code{minDeltaTime} minutes
after the last independent record of the same species
(\code{"lastIndependentRecord"}), or \code{minDeltaTime} minutes after the
last record (\code{"lastRecord"}).}

\item{timeZone}{character. Must be an argument of
\code{\link[base]{OlsonNames}}.}

\item{stationCol}{character. Name of the camera trap station column. Assumes
"Station" if undefined.}

\item{writecsv}{logical. If \code{TRUE}, a .csv file of the record table will
be saved.}

\item{outDir}{character. Directory to save .csv file to. If NULL and
\code{writecsv = TRUE}, recordTable will be written to \code{inDir}.}

\item{metadataHierarchyDelimitor}{character. The character (either "|" or ":")
that delimits hierarchy levels of image metadata tags contained in the image
management software field called "HierarchicalSubject".}

\item{metadataSpeciesTag}{character. In custom image metadata, the species ID
tag name. Only one species name can be provided (e.g. "dog").}

\item{metadataIDTag}{character. In custom image metadata, the individual ID
tag name.}

\item{additionalMetadataTags}{character. Additional camera model-specific
metadata tags to be extracted. If possible, specify tag groups as returned
by \code{\link[camtrapR]{exifTagNames}}, (e.g.
c("MakerNotes:AmbientTemperature","MakerNotes:MoonPhase").}

\item{removeDuplicateRecords}{logical. If \code{TRUE}, only one record will be
shown in the record table when there are several records of the same species
at the same station (also same camera if cameraID is defined) at exactly the
same time.}

\item{stationIDposition}{integer. The numeric position within the
filepath/directory structure of the 'Station' directories (e.g.
"C:/Documents/CamProject/Region/Site/Transect/Station/Camera/Species/...",
the 'Station' directory position = 6).Only need to use this argument when
information from one or more directories (i.e. folders) before the 'Station'
level directory are to be used to populate columns within the record table
(e.g. "SurveyPeriod/Region/Site/Transect/Station/..."). If this argument is
missing, the function assumes the 'Station' directory is next in the
filepath as specified in \code{inDir}.}

\item{speciesPosition}{integer. The numeric position within the
filepath/directory structure of the 'Species' directories (e.g.
"C:/Documents/CamProject/Region/Site/Transect/Station/Camera/Species/...",
the species directory position = 8). Only need to use this argument when
\code{speciesIDfrom} = "directory" and there are one or more directories
(i.e. folders) after the 'Species' level directory (e.g.
"SurveyPeriod/Region/Site/Transect/Station/Species/Counts/...").}

\item{cameraIDposition}{integer. The numeric position of the 'Camera'
directories within the filepath/directory structure (e.g.
"C:/Documents/CamProject/Region/Site/Transect/Station/Camera/Species,...",
the camera ID directory position = 7). Only need to use this argument when
\code{cameraIDfrom} = "directory" and there are one or more directories
(i.e. folders) after the 'Species' level directory (e.g.
"SurveyPeriod/Region/Site/Transect/Station/Species/Counts/...").}

\item{directoryInfoPositions}{integer. Vector of the numeric positions within
the filepath/directory structure of the directories containing extra
information that the user wishes to have itemised in columns within the
record table. Can take a vector of length >1. This argument must be used if
the directory filepath contains extra directories beyond 'Species' (e.g. the
directories "Site", "Transect" and "Counts" within the filepath of
"C:/Documents/CamProject/Site/Transect/Station/Camera/Species/Counts/...",
would be specified as c(3:4, 8)).}

\item{directoryInfoNames}{character. The names of the directories for the
additional information coming from the directory as specified in
\code{directoryInfoPositions}. If the length is >1, then the names need to
be concatonated and listed in the same order as the numbers specified in
\code{directoryInfoPositions} (e.g. the \code{directoryInfoPositions} of
c(3:4, 8) within the filepath of
"C:/Documents/CamProject/Site/Transect/Station/Camera/Species/Counts/...",
would be specified as c("Site", "Transect","Counts")). This argument must be
used if the directory filepath contains extra directories beyond 'Species'.}
}
\description{
The function generates a single-species record table containing individual
IDs, e.g. for (spatial) capture-recapture analyses. It prepares input for
the function \code{\link[camtrapR]{spatialDetectionHistory}}.
}
\details{
The function can handle a number of different ways of storing images.
In every case, images need to be stored in a species directory first
(e.g. using function \code{\link[camtrapR]{getSpeciesImages}}). Station subdirectories
are optional.

As with species identification, individuals can be identified in 2 different
ways: by moving images into individual directories
("Species/Station/Individual/XY.JPG" or "Species/Individual/XY.JPG") or by
metadata tagging (without the need for individual directories:
"Species/XY.JPG" or "Species/Station/XY.JPG").

\code{minDeltaTime} is a criterion for temporal independence of records of
an individual at the same station/location. Setting it to 0 will make the
function return all records. \code{camerasIndependent} defines if the
cameras at a station are to be considered independent (e.g. \code{FALSE} if
both cameras face each other and possibly \code{TRUE} if they face different
trails). \code{stationCol} is the station column name to be used in the
resulting table. Station IDs are read from the station directory names if
\code{hasStationFolders = TRUE}. Otherwise, the function will try to extract
station IDs from the image filenames (requires images renamed with
\code{\link{imageRename}}.

If individual IDs were assigned with image metadata tags,
\code{metadataIDTag} must be set to the name of the metadata tag group used
for individual identification. \code{metadataHierarchyDelimitor} is "|" for
images tagged in DigiKam and images tagged in Adobe Bridge/ Lightroom with
the default settings. Manufacturer-specific Exif metadata tags such as
"AmbientTemperature" or "MoonPhase" can be extracted if specified in
\code{additionalMetadataTags}. Multiple names can be specified as a
character vector as: \code{c(Tag1, Tag2, ...)}. Because they are not
standardized, function \code{\link[camtrapR]{exifTagNames}} provides a vector of all
available tag names. The metadata tags thus extracted may be used as
individual covariates in spatial capture-recapture models.

\code{eventSummaryColumn} and \code{eventSummaryFunction} can be used to
extract summary statistics for independent sampling events. See documentations
for these functions in \code{camtrapR}.
}
\examples{


\dontrun{   # the examples run too long to pass CRAN tests

 wd_images_ID_individual <- system.file("pictures/sample_images_indiv_tag/LeopardCat",
                                        package = "camtrapR")
 # missing space in species = "LeopardCat" is because of CRAN package policies
 # note argument additionalMetadataTags: contains tag names as returned by function exifTagNames

 if (Sys.which("exiftool") != ""){        # only run these examples if ExifTool is available

 rec_table_pbe <- recordTableIndividualAssist(inDir                  = wd_images_ID_individual,
                                        minDeltaTime           = 60,
                                        deltaTimeComparedTo    = "lastRecord",
                                        StationIDfrom          = "filename",
                                        individualIDfrom       = "metadata",
                                        camerasIndependent     = FALSE,
                                        writecsv               = FALSE,
                                        metadataIDTag          = "individual",
                                        additionalMetadataTags = c("EXIF:Model", "EXIF:Make"),
                                        timeZone               = "Asia/Kuala_Lumpur"
                                        speciesIDfrom          = "directory"
 )



} else {
# show function output if ExifTool is not available
message("ExifTool is not available. Cannot test function. Loading recordTableSample instead")
data(recordTableSample)
}
}


}
\references{
Phil Harvey's ExifTool \url{https://exiftool.org/}
}
\author{
Juergen Niedballa
}
