% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variousOtherHelperFunctions.R
\name{addStationCameraID}
\alias{addStationCameraID}
\title{Add station camera ID}
\usage{
addStationCameraID(
  intable,
  dirs_short,
  stationCol,
  cameraCol,
  cameraID,
  hasStationFolders,
  i,
  IDfrom,
  stationIDposition,
  cameraIDposition = NULL
)
}
\arguments{
\item{intable}{data.frame or data.table. A data frame or a data.table with the
detections (e.g. the output from recordTable)}

\item{dirs_short}{Vector or directory names (without path)}

\item{stationCol}{character. Name of the camera trap station column. Assumes
"Station" if undefined.}

\item{cameraCol}{Character. The name of the camera column (e.g. "Camera")}

\item{cameraID}{This argument is deprecated. Use \code{cameraIDfrom}.}

\item{hasStationFolders}{Logical. Whether station column is taken from directory}

\item{i}{Position index taken from the for loop}

\item{IDfrom}{This argument is deprecated. Use \code{speciesIDfrom}.}

\item{stationIDposition}{integer. The numeric position within the
filepath/directory structure of the 'Station' directories (e.g.
"C:/Documents/CamProject/Region/Site/Transect/Station/Camera/Species/...",
the 'Station' directory position = 6).Only need to use this argument when
information from one or more directories (i.e. folders) before the 'Station'
level directory are to be used to populate columns within the record table
(e.g. "SurveyPeriod/Region/Site/Transect/Station/..."). If this argument is
missing, the function assumes the 'Station' directory is next in the
filepath as specified in \code{inDir}.}

\item{cameraIDposition}{integer. The numeric position of the 'Camera'
directories within the filepath/directory structure (e.g.
"C:/Documents/CamProject/Region/Site/Transect/Station/Camera/Species,...",
the camera ID directory position = 7). Only need to use this argument when
\code{cameraIDfrom} = "directory" and there are one or more directories
(i.e. folders) after the 'Species' level directory (e.g.
"SurveyPeriod/Region/Site/Transect/Station/Species/Counts/...").}
}
\description{
Internally used to compile record table.
}
\details{
Do not need to update cameraID and IDfrom argument because these are provided
internally with correct argument and the function is not exported
}
